Fix and simplify page change calculations
Fix the page manipulations when start scrolling to the (for exampl)
left page and then scroll to the right one, not releasing the touch.
Or when start scrolling but then release the page so that it returned
to the initial position.

No more nonobvious "always keep right page" logic. Now when IDLE there
is only one page added: the current page. The secondary page is added
when moving to the left or to the right from the primary page. The
(current) secondary page is removed when changing scroll state to IDLE.